pkvm: x86: init: check if pkvm IOMMU can support all the enumerated PCI devices
The PCI device count determines how much memory will be consumed by
pkvm IOMMU driver to shadow the root/context/pasid page tables at the
run time. As pkvm's memory is reserved at the boot time, the PCI device
count needs to be pre-defined. Later patch will reserve the memory for
IOMMU according to this number. During pkvm initialization, need to check
if the enumerated PCI device count on this platform beyond this number or
not.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>