pkvm: x86: not to handle EPT violation for GPA overlapped with IOMMU
IOMMU is owned by pkvm, and if the GPA caused EPT violation is in the
IOMMU MMIO range, pkvm should not create mapping for it. So before create
the EPT mapping, do the check to guarantee this.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
(cherry picked from commit 57239a468471ef797ed22b6b23b5d5bea98888fc)
Signed-off-by: Vineeth Pillai <vineethrp@google.com>