pkvm: x86: intrdoucing IN/OUTSIDE_GUEST_MODE vcpu mode
With adding kick vcpu support, vcpu can handle some pending requests
before vmenter to non-root mode.

Using vcpu mode to indicate if a vcpu should be kicked or not. Setting
the vcpu mode to be IN_GUEST_MODE when it is going to be in non-root
mode, and changing to OUT_GUEST_MODE when it exits from non-root mode.
And only kick vcpu when it is IN_GUEST_MODE. So vcpu will handle the
pending requests when it is in OUT_GUEST_MODE and after that changing to
IN_GUEST_MODE.

It is possible that some kick is lost due to the mode changing so need
to double check if there is any pending requests after changing to
IN_GUEST_MODE. This will be introduced by later patch when is really
making any requests.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Reviewed-by: Jason Chen CJ <jason.cj.chen@intel.com>
(cherry picked from commit 65abee5a60473767fe0f0eb3f34cbd3b711dffe4)
Signed-off-by: Vineeth Pillai <vineethrp@google.com>