pkvm: x86: init: isolate IOMMU from host VM
As IOMMU is maintained by pkvm, the IOMMU MMIO mapping in EPT needs
to be removed to isolate the IOMMU from host VM, so that the IOMMU MMIO
accessing from the host IOMMU driver can be captured by pkvm and
emulate. The isolation is done in the finalise hypercall.

Normally, after remove the MMIO accessing from EPT, such access will
trigger EPT violation vmexit then doing instruction & IO emulation in
the hypervisor. While we don't expect to do instruction emulation in
pkvm hypervisor, so to simplify the IOMMU MMIO vmexit, introduced the
pkvm_read/write helpper functions, which relies on pkvm_enabled flag to
determine the way of accessing the MMIO. If pkvm_enabled is true,
uses the hypercall to access MMIO. And if pkvm_enabled is false, fallback
to the traditional way to do the read and write.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>