pkvm: x86: add spin lock to protect MMU map/unmap
pkvm MMU map/unmap will be used at the runtime when emulating
IA32_APIC_BASE msr for primary VM which will be introduced by the nest
patch. So it is possible that multiple CPUs can do the mmu map at the
same time. Introduce a spin lock to avoid the possible race.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Reviewed-by: Jason Chen CJ <jason.cj.chen@intel.com>