pkvm: x86: Implement do_unshare() helper for unsharing memory
do_unshare() is paired with do_share(), and it's the reverse of the
do_share(), which results in the page sharer take back accessibility of
the shared page for another component which being shared. The page state
become like:

sharer: PAGE_SHARED_OWNED       => OWNED
be shared: PAGE_SHARED_BORROWED => NOPAGE

Introduce a do_unshare() helper for safely unsharing a memory region
from one component to another. It will check the page state stored in
pte, and make sure the do_unshare() is paired with do_share(), this will
make the page being correctly take back and can only be accessed by the
page owner after unsharing.

Signed-off-by: Shaoqin Huang <shaoqin.huang@intel.com>