pkvm: x86: add pkvm_pgtable_unmap_safe for a safe unmap
The current pkvm_pgtable_unmap API takes a phys parameter which will
check if the unmap entry really maps this phys memory. This is not
necessary for all the unmap cases, and actually phys is unknown to some
cases.

However, for some cases, still prefer to perform such check to avoid
unmapping an incorrect entry because such failure is not easy to be
detected but cause a serial security issue. The unmap from page state
API and from the host EPT are the cases better to have such check. So
introduce a safe version unmap API.

The safe version unmap API will be used by page state machine and host
EPT to do unmap. For the MMU page table, it can still use the normal
version of the unmap.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>