pkvm: x86: use pgtable unmap API to emulate invept
Currently emulating invept is by using destroy API which can unmap the
entire shadow EPT with also freeing the root page. To avoid freeing root
page, needs to hold additional reference count for this. Actually this
is excatly what unmap API does, which unmap the pgtable with a range
while keeping the root page not freed. So switch to use unmap API.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Reviewed-by: Jason Chen CJ <jason.cj.chen@intel.com>