pkvm: x86: Add new hypercall for KVM high to share passthrough dev info
pKVM needs to know the passthrough device information for a protected VM
so that can set up the pgstate pgt with proper IOMMU translations. Such
information is shared through hypercall when KVM high is adding a
vfio-group. All the devices' BDF in the same IOMMU group will be send
to pKVM.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Reviewed-by: Jason Chen CJ <jason.cj.chen@intel.com>