pkvm: x86: add addr and size parameter to flush_tlb callback
pkvm IOMMU driver re-use the host EPT as the second-stage page table
with both scalable IOMMU and legacy mode IOMMU for some specific
configurations. So each time when some memory is unmapped from host
EPT, not only needs to flush the EPT tlb but also needs to flush the
IOMMU iotlb. pkvm can resue the flush_tlb callback for this purpose.

As the iotlb flushing can be page selective with the addr and size
information if IOMMU can support, add parameters in flush_tlb callback
to pass this information in.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Reviewed-by: Jason Chen CJ <jason.cj.chen@intel.com>