pkvm: x86: hide EPT A/D support from pKVM high
The Accessed and Dirty bits in virtual EPT which is maintained by L1
KVM high don't reflect the actual access status of L2 GPAs.  If KVM
uses these bits as indications of whether a GFN can be swapped out,
it could cause problems.

If we hide the A/D bits from KVM, it may correctly emulate the EPT A/D
bits with the access tracking faults facility.

Signed-off-by: Zide Chen <zide.chen@intel.com>