KVM: x86: Using SET_MMIO_VE hypercall to clear "Suppress #VE" bit
In KVM, if the gpa doesn't have a memslot, that means a mmio pte will
be set on the gpa. And KVM will set a mmio pte which will cause
misconfiguration. But for protected VM, we want to set a pte which can
trigger #VE, so we use the SET_MMIO_VE hypercall to clear the
"Suppress #VE" bit and set the pte as invalid in shadow ept in pkvm.

After setting this mmio pte, the next time the protected VM accesses
mmio, the protected VM will receive a virtualzation exception. And the
flow to handle this #VE is at below:

1. Protected VM receives #VE, jump into the #VE handler.
2. In #VE handler, protected VM uses GET_VE_INFO hypercall to get #VE
   information.
3. Protected VM uses #VE information to do instruction decode.
4. Protected VM uses IOREAD/IOWRITE hypercall to pass the MMIO
   information (include gpa, size, val) to do mmio emulation in KVM.
5. KVM return the emulation results to protected VM. Protected VM
   saves the mmio results into register and return from the #VE handler.
6. Protected VM returns back to the mmio instruction, and get the
   result. Continue to run.

This patch only supports EPT_VIOLATION_VE for tdp_mmu, and make sure
tdp_mmu_enabled is true when using pkvm.

Signed-off-by: Shaoqin Huang <shaoqin.huang@intel.com>
Reviewed-by: Jason Chen CJ <jason.cj.chen@intel.com>

dmaluka@
1. Note that TDP MMU is currently temporarily disabled in ChromeOS:
   http://crrev/c/5152509 so when testing pKVM-IA in ChromeOS, make sure
   to enable it by removing kvm.tdp_mmu=0 from the kernel command line,
   to make MMIO work in Linux pVMs.
2. Apparently this solution is not fully working: it is observed that a
   Linux pVM access to some of its MMIO addresses (namely to 0xF4000000
   which is its guest PCI MMCFG base address) doesn't work, since
   fault->slot is not NULL for some reason, so pkvm_set_mmio_ve() is not
   called.

Signed-off-by: Dmytro Maluka <dmaluka@google.com>
(cherry picked from commit 18117ea312f0bf4c917f944eebe6213f7fc59a98)
Signed-off-by: Vineeth Pillai <vineethrp@google.com>