pkvm: x86: Initialize every SEPT entry with "suppress #VE" bit set
The KVM MMIO handler in primary VM can't work for protected VM (pVM)
anymore, as KVM needs to access pVM's memory for instruction decoding,
while the memory of pVM is protected by pKVM.

In this situation, we expect pVM to do instruction decoding by itself,
then explicitly make hypercall to KVM for MMIO emulation. So we need EPT
violation caused by MMIO access to be routed to pVM.

Once EPT_VIOLATION_VE is enabled and the "suppress #VE" bit (which is
bit 63) in a EPT PTE is clear, the pVM will receive a #VE instead of a
EPT violation when accessing the corresponding page. pKVM can leverage
this mechanism by configuring the MMIO EPT PTE to trigger #VE for pVM.

This patch is preparing to enable EPT_VIOLATION_VE, before enabling it,
all the EPT PTE shall set "suppress #VE" bit by default which leads to
changes for operations of PTE allocation, unmap etc.

Signed-off-by: Shaoqin Huang <shaoqin.huang@intel.com>
Reviewed-by: Jason Chen CJ <jason.cj.chen@intel.com>