pkvm: x86: Introduce pkvm_ptdev structure
In pkvm, each PCI device has an instance of pkvm_ptdev structure to track
device specific information, like the did, pgt. The pkvm_ptdev instance
will be created and linked to its corresponding IOMMU if it cannot be
find from the ptdev list of this IOMMU when creating its shadow pasid
entry. The ptdev instance will be released when IOMMU clearing its pasid
entry.

The pkvm_ptdev instance will also be used by pKVM to isolate the
passthrough devices of a protected VM.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Reviewed-by: Jason Chen CJ <jason.cj.chen@intel.com>