pkvm: x86: Add support of cpuid for protected VM in pkvm
Protected VM can use cpuid to detect if it is running on pkvm
hypervisor, so add support for emulating cpuid for protected VM in pkvm.

Since TDX using leaf=0x21, subleaf=0x0 to detect if guest running on
TDX. We reuse the same cpuid to detect if guest running on pkvm, but
return different value in register.

Signed-off-by: Shaoqin Huang <shaoqin.huang@intel.com>
Reviewed-by: Jason Chen CJ <jason.cj.chen@intel.com>