pkvm: x86: Add pkvm guest virtualization exception (#VE)
This patch includes the guest support for the virtualization exception.
When guest do the MMIO operation, guest will receive #VE and jump to the
exception handler.

In the #VE handler, guest will first use hypercall to get #VE
information, and then it will do instruction decode by itself. And using
hypercall to pass the mmio information to KVM to emulate it.

After KVM emulate this MMIO, guest will return from the #VE handler.
After it return, guest execute the MMIO operation success, and continue
to run.

Signed-off-by: Shaoqin Huang <shaoqin.huang@intel.com>
Reviewed-by: Jason Chen CJ <jason.cj.chen@intel.com>