REVERTME: kvm: not to pin the non-memory pfn
With the passthrough device support in protected VM, passthrough device
MMIO bar will be mapped in EPT. For such non-memory pfn, should not to
pin the page.

As this patch is in addition to the REVERTME commit
"d03dca637960a5f3b5ada0b01cda64596758c18b", also mark with REVERTME so
that they can be reverted together once the private fd solution is ready.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Reviewed-by: Jason Chen CJ <jason.cj.chen@intel.com>