pkvm: x86: Adjust legacy IOMMU mode to use ptdev framework
Reorganize the code to properly register and use ptdev structures for
passthrough devices in legacy mode just like in scalable mode.

This patch does not change any functionality yet but it prepares for
the future patches that will add support for attaching passthrough
device to protected VM in legacy mode and for shadowing IOMMU page
tables for memory protection in legacy mode.

Signed-off-by: Dmytro Maluka <dmy@semihalf.com>