pkvm: x86: Introduce iommu_coherency flag for ptdev
This flag represents the coherency of the IOMMU unit this ptdev is on.
It is configured when allocate a new ptdev, either during syncing IOMMU
pasid/context or attaching a ptdev to a shadow VM. So a new interface is
added in pKVM IOMMU driver to get the IOMMU coherency by bdf/pasid when
attaching a ptdev. The current implementation does ptdev "get" and
"allocate"(if not existing) in one single function pkvm_get_ptdev(). As
ptdev "get" doesn't need to configure iommu_coherency and only "allocate"
does, separate the "allocate" from pkvm_get_ptdev().

The iommu_coherency flag is used by the next patch to configure the
mm_ops for this ptdev's IOMMU page table.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Reviewed-by: Jason Chen CJ <jason.cj.chen@intel.com>