pkvm: x86: Fix unitialized iotlb_lm_invalidate() return value
When the host requests DMA_TLB_DSI_FLUSH or DMA_TLB_PSI_FLUSH iotlb
flush for the given DID, it is possible that there is no device with
this DID set up yet. For instance, this happens (transiently) when
starting a VM with a passthrough device. This is not a faulty situation,
so in such case iotlb_lm_invalidate() is essentially a no-op and should
return success, i.e. zero. So make sure that it returns zero in this
case, not an undefined uninitialized value.

Without this fix, with CONFIG_PKVM_INTEL_DEBUG disabled, when we start a
VM with a passthrough device, the VM gets stuck consuming 100% CPU time
(and cannot be killed), due to pKVM continuously receiving and handling
hypercalls from the host IOMMU driver. It is caused by
iotlb_lm_invalidate() constantly returning random non-zero values, which
results in pKVM not setting the invalidation descriptor status as
successful (QI_DONE), which results in the host IOMMU driver's
qi_submit_sync() endlessly spinning in a loop waiting until the
invalidation status becomes QI_DONE. In other words, the transient
condition mentioned aboved is not transient anymore.

Interestingly, with CONFIG_PKVM_INTEL_DEBUG enabled this issue has not
been observed so far. Due to some subtle differences in behavior, with
CONFIG_PKVM_INTEL_DEBUG enabled, once in 10-20 times the undefined value
returned by iotlb_lm_invalidate() is luckily zero, so we are not getting
stuck in the described condition endlessly.

Fixes: c6bbb5885bd8 ("pkvm: x86: Use shadow IOMMU page tables for legacy IOMMU")
Signed-off-by: Dmytro Maluka <dmaluka@google.com>