pkvm: x86: De-hardcode vm_type size in __pkvm_init_shadow_vm()
The size of the vm_type field is not guaranteed to remain the same.
For instance, recently in commit 517987e3fb19 ("KVM: x86: add fields to
struct kvm_arch for CoCo features"), the vm_type type was changed from
unsigned long to u8, so __pkvm_init_shadow_vm() needed to be updated
accordingly. So de-hardcode it now, to avoid wasting time on debugging
if it changes again in the future.

Signed-off-by: Dmytro Maluka <dmaluka@google.com>