pkvm: x86: Keep the __pkvm_ prefix for all pkvm symbols in debug mode
The previous implementation is to remove the "__pkvm_" prefix in debug
mode so that can reuse the linux kernel printk symbols. The side effects
of doing this are: 1) not able to detect if pkvm has involved any other
linux kernel symbols or not if develop in debug mode. 2) Not to able
verify the libraries implemented by pkvm as the pkvm's implementation
should be disabled due to the conflict of library symbols. Especially
the 2nd side effact will cause troubles for code sharing.

To resolve these, introduced a debug.c which will be compiled as part
of the linux kernel code (not pkvm hypervisor code) for debug mode only.
As the debug mode also adds the "__pkvm_" prefix to the required linux
kernel printk symbols, the debug.c implements a trampoline for this.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>