pkvm: vmx: Enable EPT violation VE according to vmcs_config in pkvm_hyp
Although the current implementation for pVM requires to use #VE to
decode the MMIO instructions, add the EPT VE as an opt setting as if
make it as min setting, then pkvm may not be able to run on the platform
which doesn't have VE capability. This is not friendly to the POC
development. And eventually the pVM doesn't need to use #VE to decode as
pkvm hypervisor will do so in the future. So make it as opt and only
enable EPT VE for the nested pVM when the vmcs config supports this.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>