pkvm: x86: Allocate memory page for percpu
Allocate memory pages for the percpu pages are from the reserved memory
pool, which will be protected after host being deprivileged. The number
of required pages will be calculated according the pkvm hypervisor
percpu section size, and the percpu section offset will also be used to
calculate the accurate percpu offset by the following patch.

Considerring that, pkvm hypervisor will have architecture in-dependent
code, e.g., the percpu support, which could be possible benifit SEV as
well, reusing arch/x86/kvm/vmx/pkvm seems doesn't make sense. So created
a new folder at arch/x86/kvm/pkvm to hold such kind of code. Based on
this, the architecture specific code may be better to put together under
this new folder, with the tree hierarchy like this:
arch/x86/kvm/pkvm:
 |--xxx.c
 |--vmx
 |  |--yyy.c
 |  `--
 |--sev
 |  |--zzz.c
 |  `--
The architecture independent code can be put directly in the pkvm/. The
VMX and SEV specific code can be put in the pkvm/vmx and pkvm/sev
respectively.

However this might not be the final perferred hierarchy so keeping the
existing pkvm code where they are right now, but will put the new
introduced pkvm code in this new location during the development in the
future, to see how this will look like.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>