kvm: pkvm: Define kvm_call_pkvm helper
To support the PV ABI, defines a helper function kvm_call_pkvm for linux
kvm to call the PV methods in the pkvm hypervisor by using KVM_CALL
hypercall. To save the hypercall number, it will require at least two
input parameters. The first one is the vmcall number and the second is
the method number which represents the required method. Besides these
two inputs, maximal three additional input parameters can be supported
which work as the inputs for the required method. The helper can be
extended to support more parameters if a pkvm method requires in the
future.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>