pkvm: vmx: Add struct pkvm_vcpu_vmx
Struct pkvm_vcpu_vmx represents a vcpu structure owned by pkvm
hypervisor for VMX. It requires struct kvm_vcpu sitting at the offset 0
so that it would be appended in the end of pkvm_vcpu. This is achaived
by putting the struct vcpu_vmx at the first field which wraps the struct
kvm_vcpu at the offset 0.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>