pkvm: vmx: Reuse vmx instructions defined in vmx/vmx_ops.h
To reuse the vmx/vmx_ops.h instead of redefining one for pkvm, some
symbols for handling vmx instruction errors should be implemented in
pkvm. These symbols are ported from the files in the folder of kvm/vmx/
to the kvm/pkvm/vmx/, with keeping the same file name and the same code
logic. Because some of these code are annotated with noinstr, keep the
noinstr attribute and extend the pkvm linker script to include
.noinstr.text as part of pkvm text.

For non-debug mode, undefine CONFIG_GENERIC_BUG/CONFIG_BUG to bypass
the panic logic implemented in linux kernel so the error handling will
basically do nothing for non-debug mode. For debug mode, implement the
symbols which requires the prefix "__pkvm_".

With this, the pkvm vmx_ops.h can be removed.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>