pkvm: x86: Introduce uAPI to set/query pvmfw info
Introduce uAPI to let the VMM discover the pvmfw region size and specify
the address where to load pvmfw in the guest memory, very similar to the
uAPI implemented in pKVM on ARM [1].

This uAPI looks strange, and it is temporary, to be replaced with a
proper uAPI when upstreaming this part of pKVM code in the future. But
for now, let's keep the x86 uAPI very similar to the existing ARM one,
to avoid unneeded discrepancies from pKVM-ARM (since the functionality
provided by this uAPI is exactly the same on both architectures).

[1] https://android.googlesource.com/kernel/common/+/39e66da69481f1ba70d5cfb6cdf23ba07e2377ef%5E%21/

Signed-off-by: Dmytro Maluka <dmaluka@google.com>